Give up on static linking with homebrew
When autobrew is off. It does not work now on macOS 14,
because pkg-config adds -lzstd, but the linker does not
find that, unless we also include zstd in the pkg-config
command lines.

Plus you might end up with dynamic linking anyway.
So it is just easier to do dynamic linking, and rely on
autobrew for CRAN.